requirements.txt
streamlit==1.8.1
torch==1.10.0
Pillow
matplotlib
